backend/requirements.txt
requests==2.26.0
Django==3.2.25
djangorestframework==3.12.4
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-filter==2.3.0
psycopg2-binary==2.9.3 
gunicorn==20.1.0 
Pillow==9.0.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
pytils==0.4.1
python-dotenv==1.0.1